+++ pkg/operator/operands/admission/resources.go
@@ -365,5 +365,9 @@ func buildArgsList(kaiConfig *kaiv1.Config, config *kaiv1admission.Admission) []
 
 	common.AddK8sClientConfigToArgs(config.Service.K8sClientConfig, args)
 
+	if kaiConfig.Spec.Global.JSONLog != nil && *kaiConfig.Spec.Global.JSONLog {
+		args = append(args, "--zap-devel=false")
+	}
+
 	return args
 }