+++ pkg/operator/operands/pod_grouper/resources.go
@@ -91,6 +91,10 @@ func buildArgsList(kaiConfig *kaiv1.Config) []string {
 		args = append(args, "--pod-label-selector", formatLabelSelector(kaiConfig.Spec.Global.PodLabelSelector))
 	}
 
+	if kaiConfig.Spec.Global.JSONLog != nil && *kaiConfig.Spec.Global.JSONLog {
+		args = append(args, "--zap-devel=false")
+	}
+
 	return args
 }